requirements.txt
Django==1.7.6
django-classy-tags==0.5.1
django-cms==3.0.6
django-mptt==0.6.1
django-nose==1.3
django-reversion==1.8.5
django-sekizai==0.7
djangocms-admin-style==0.2.2
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.1
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.5.4
djangocms-link==1.5
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.0
djangocms-video==0.1
MySQL-python==1.2.3
psycopg2==2.6.1
South==1.0
django-decorator-include==0.1
django-render-partial==0.1
django-ckeditor==5.0.3
